Collecting Bill Information from a scrapeBill Object.

Collect Bill Information and Put Into Data Frame.

collectBillData(x, include = c("BillID", "URL", "Title", "Summary",
  "Sponsorship", "Action", "Committee", "Subject"), expand = "None",
  progress = FALSE)

Arguments

x
A (or A list of) scrapeBill object created by scrapeBill function.

include
character vector of What element(s) of bill information to include in output. Options are "BillID", "URL", "Title", "Summary", "Sponsorship", "Action", "Committee", "Subject", and "Text". The default is to include everything except for "Text". Need at least one element. Information set by this option uses single row output (i.e., Set officialonly, latestonly, single, and allinone to TRUE in collectOPTION functions).

expand
The default is "None" (not to expand data). Else, create long data by using expanded information (i.e., Set officialonly, latestonly, single, and allinone to FALSE in collectOPTION functions). Choose option from ONE of "Title", "Summary", "Sponsorship", "Action", "Committee", "Subject", and "Text". CANNOT set more than one option.

progress
If TRUE, show progress bar for the processing of multiple bills.

Value

A data.frame object of bill information. See functions in See Also section for variable details.

See also

Examples

tgturl <- "https://www.congress.gov/bill/116th-congress/senate-bill/252"
bill252 <- scrapeBill(tgturl)
collectBillData(bill252)

#>          ID Congress Chamber Type Number
#> 1 S11600252      116  Senate Bill    252
#>                                                                                                          Name
#> 1 A bill to authorize the honorary appointment of Robert J. Dole to the grade of colonel in the regular Army.
#>                                                            URL
#> 1 https://www.congress.gov/bill/116th-congress/senate-bill/252
#>                                                                                                         Title
#> 1 A bill to authorize the honorary appointment of Robert J. Dole to the grade of colonel in the regular Army.
#>                                                                                                                      Summary
#> 1 This bill provides for the honorary appointment of Robert J. Dole, of Kansas, to the grade of colonel in the regular Army.
#>   Sp_Chamber      Sp_Name Sp_LastName Sp_Party Sp_State Sp_District Cosp_N
#> 1     Senate Roberts, Pat     Roberts        R       KS        <NA>      2
#>   Cosp_Original Cosp_D Cosp_R ActLatestDate_Overview ActLatestText_Overview
#> 1             1      0      1             03/04/2019       Passed in Senate
#>   ActLatestDate_All                                  ActLatestText_All
#> 1        03/05/2019 Referred to the House Committee on Armed Services.
#>   VoteTimes_House VoteLatestDate_House VoteLatestType_House
#> 1               0                 <NA>                 <NA>
#>   VoteLatestText_House VoteTimes_Senate VoteLatestDate_Senate
#> 1                 <NA>                0                  <NA>
#>   VoteLatestType_Senate VoteLatestText_Senate VoteTimes_House.1
#> 1                  <NA>                  <NA>                 0
#>   RCLatestDate_House RCLatestText_House RCLatestNumber_House RCLatestYea_House
#> 1               <NA>               <NA>                   NA                NA
#>   RCLatestNay_House RCTimes_Senate RCLatestDate_Senate RCLatestText_Senate
#> 1                NA              0                <NA>                <NA>
#>   RCLatestNumber_Senate RCLatestYea_Senate RCLatestNay_Senate Committee_N
#> 1                    NA                 NA                 NA           2
#>                               Committee_Names     Subject_Primary Subject_N
#> 1 Senate Armed Services; House Armed Services Private Legislation         1
#>           Subject_All
#> 1 Private Legislation