Collecting Sponsorship Information from a scrapeBill Object.

Collect information regarding sponsorship of the bill.

collectSponsorship(x, single = TRUE, progress = FALSE)

Arguments

x
A (or A list of) scrapeBill object created by scrapeBill function.

single
Whether to export single row results (see following description). The default is TRUE.

progress
If TRUE, show progress bar for the processing of multiple bills.

Value

If single==TRUE, returns data.frame object with following variables (single row):

Sp_Chamber: Chamber of Primary Sponsor

Sp_Name: Full Name of Primary Sponsor

Sp_LastName: Last Name of Primary Sponsor

Sp_Party: Party of Primary Sponsor

Sp_State: State of Primary Sponsor

Sp_District: District of Primary Sponsor (Only for House Representatives)

Cosp_N: Total Number of Cosponsors

Cosp_Original: Number of Original Cosponsors

Cosp_D: Number of Democratic Cosponsors

Cosp_R: Number of Republican Cosponsors

If single==FALSE, data.frame object with following variables (rows represent sponsors):

Date: Date of Sponsorship. If Primary Sponsor, this is the bill introduction date.

Status: Status of Sponsorship. One of "Primary Sponsor," "Original Cosponsor," and "Cosponsor."

Chamber: Chamber of Sponsor

Name: Full Name of Sponsor

LastName: Last Name of Sponsor

Party: Party of Sponsor

State: State of Sponsor

District: District of Sponsor (Only for House Representatives)

See also

Examples

tgturl <- "https://www.congress.gov/bill/109th-congress/house-bill/19"
bill19 <- scrapeBill(tgturl)
collectSponsorship(bill19)

#>   Sp_Chamber      Sp_Name Sp_LastName Sp_Party Sp_State Sp_District Cosp_N
#> 1      House Calvert, Ken     Calvert        R       CA          44     37
#>   Cosp_Original Cosp_D Cosp_R
#> 1            12      0     36