Collecting Text Information from a scrapeBill Object.

Collect information regarding the latest full text of the bill.

collectText(x, allinone = TRUE, progress = FALSE)

Arguments

x
A (or A list of) scrapeBill object created by scrapeBill function.

allinone
If TRUE (default), the function returns all texts of bill into one character scalar. If FALSE, the function separates text by potential sections and exported as list.

progress
If TRUE, show progress bar for the processing of multiple bills.

Value

A character or list object of Bill Texts.

See also

Examples

tgturl <- "https://www.congress.gov/bill/116th-congress/senate-bill/252"
bill252 <- scrapeBill(tgturl)
collectText(bill252)

#> [1] "116th CONGRESS1st Session S. 252 IN THE HOUSE OF REPRESENTATIVES March 5, 2019 Referred to the Committee on Armed Services AN ACT To authorize the honorary appointment of Robert J. Dole to the grade of colonel in the regular Army. Be it enacted by the Senate and House of Representatives of the United States of America in Congress assembled, SECTION 1. Findings. Congress makes the following findings: (1) Robert J. Dole, also known as Bob Dole, was born July 22, 1923, in Russell, Kansas. (2) As a student at the University of Kansas, Bob Dole enrolled in the Army Enlisted Reserve Corps in 1942. He was called to active duty the following year and served during World War II. While deployed to Italy as an infantry lieutenant in the 10th Mountain Division, he was seriously wounded in combat and was twice cited for acts of heroism under fire. He finished his military service with two Purple Hearts and two awards of the Bronze Star Medal with “V” device for valor. He was also awarded the American Campaign medal, the European-African-Middle Eastern Campaign Medal, and the World War II Victory medal. He resigned his Army commission at the grade of captain. (3) Bob Dole served in the House of Representatives from 1961 to 1969. He served in the Senate from 1969 until 1996. His service was previously recognized by both the Legislative Branch of the United States Government, through the Congressional Gold Medal, and the Executive Branch of the United States Government, through the Presidential Medal of Freedom. (4) Separate from the actions underlying the honors previously bestowed upon him, Bob Dole also made direct contributions to the Department of Defense and the veterans of United States military service, including through the following: (A) Bob Dole sponsored or co-sponsored bills in Congress to improve veterans benefits, to improve accountability of missing members of the Armed Forces, to establish the Persian Gulf War Veterans Health Registry, to recognize women veterans, and to provide relief from certain inequities for National Guard technician service in connection with civil service retirement. (B) Bob Dole served as the National Chairman of the WWII Memorial Campaign, co-chair of the Families of Freedom Scholarship Fund, and co-chair of the Presidential Commission on the Care of America’s Returning Wounded Warriors. (5) In 2018, an Army advisory panel reviewed the comprehensive record of Bob Dole's service to the Armed Forces, veterans, and the national security of the United States. The panel, recognizing Bob Dole’s unique contributions to the Department of Defense and veterans, recommended the honorary promotion of Bob Dole to the grade of colonel in the Army. (6) Acting under provisions of section 1563 of title 10, United States Code, the Secretary of the Army endorsed the recommendation of the advisory panel for an honorary promotion of Bob Dole and forwarded it to Congress for further action. SEC. 2. Honorary appointment of Robert J. Dole to the grade of colonel in the regular Army. (a) Honorary appointment.―The honorary appointment of Robert J. Dole, of Kansas, to the grade of colonel in the regular Army is hereby authorized. (b) Additional benefits not To accrue.―The honorary appointment of Robert J. Dole to the grade of colonel in the regular Army under subsection (a) shall not affect pay or other benefits from the United States to which Robert J. Dole is otherwise entitled based upon his military service or affect any benefits to which any other person may become entitled based on his military service. Passed the Senate March 4, 2019. Attest: julie e. adams, Secretary"