Get Bill URL or HTML by Bill Number

Get Bill Suummary Page URL or scrapeBill object by Bill Number and Congress Number.

getBill(number, congress, chamber = "House", type = "Bill",
  out = "URL")

getscrapeBill(number, congress, chamber = "House", type = "Bill")

getBillDF(infoDF, out = "URL", progress = FALSE)

getscrapeBillDF(infoDF, progress = TRUE)

Arguments

number
numeric scalar of Bill number

congress
numeric scalar of Congress number

chamber
Choose from "House" (default) or "Senate"

type
Choose from "Bill" (default), "Simple Resolution", "Joint Resolution", "Concurrent Resolution", and "Amendment"

out
The output type. "URL" (default) to export bill summary page URL and "scrapeBill" to export "scrapeBill" object.

infoDF
(Only used in getBillDF function) data.frame object with columns number, congress, chamber, and type. Rows should represent individual bills.

progress
(Only used in getBillDF function) If TRUE, show progress bar for the processing of multiple bills.

Value

character scalar of URL or "scrapeBill" object.

See also

Examples

a <- getBill(1, 103)
b <- scrapeBill(a)
collectBillID(b)

#>          ID Congress Chamber Type Number                                 Name
#> 1 H10300001      103   House Bill      1 Family and Medical Leave Act of 1993

a <- getBill(1, 103, "Senate", "Joint Resolution")
b <- scrapeBill(a)
collectBillID(b)

#>            ID Congress Chamber             Type Number
#> 1 SJR10300001      103  Senate Joint Resolution      1
#>                                                                                                                                                                             Name
#> 1 A joint resolution to ensure that the compensation and other emoluments attached to the office of Secretary of the Treasury are those which were in effect on January 1, 1989.

a <- getBill(1, 101, out="scrapeBill")
collectBillID(a)

#>          ID Congress Chamber Type Number
#> 1 H10100001      101   House Bill      1
#>                                                             Name
#> 1 Department of Housing and Urban Development Reform Act of 1989

a <- data.frame(number = c(1,2,3,4),
                congress = c(103,104,108,109),
                chamber = "House",
                type = "Bill")
b <- getBillDF(a, out="scrapeBill")
collectBillID(b)

#>          ID Congress Chamber Type Number                                 Name
#> 1 H10300001      103   House Bill      1 Family and Medical Leave Act of 1993
#> 2 H10400002      104   House Bill      2                   Line Item Veto Act
#> 3 H10800003      108   House Bill      3                            Reserved.
#> 4 H10900004      109   House Bill      4       Pension Protection Act of 2006