Collecting Action Information from a scrapeBill Object.

Collect information regarding action(s) made on the bill.

collectAction(x, latestonly = TRUE, progress = FALSE)

Arguments

x
A (or A list of) scrapeBill object created by scrapeBill function.

latestonly
If TRUE (default), the function returns summary of latest actions in single row. If FALSE, the function returns data.frame object of all actions and their dates and timings.

progress
If TRUE, show progress bar for the processing of multiple bills.

Value

If latestonly==TRUE, returns data.frame object with following contents (single row):

ActLatestDate_Overview: Date of Latest Action in Actions Overview

ActLatestText_Overview: Overview Content of Latest Action in Actions Overview

ActLatestDate_All: Date of Latest Action in All Actions

ActLatestText_All: Detailed Content of Latest Action in All Actions

VoteTimes_House: Number of Voting Opportunities in House

VoteLatestDate_House: Date of Latest Vote Action in House

VoteLatestType_House: Type of Latest Vote in House

VoteLatestText_House: Content of Latest Vote in House

VoteTimes_Senate: Number of Voting Opportunities in Senate

VoteLatestDate_Senate: Date of Latest Action in Senate

VoteLatestType_Senate: Type of Latest Vote in Senate

VoteLatestText_Senate: Content of Latest Vote in Senate

RCTimes_House: Number of Roll Call Voting Opportunities in House

RCLatestDate_House: Date of Latest Roll Call in House

RCLatestText_House: Content of Latest Roll Call Vote in House

RCLatestNumber_House: ID Number of Latest Roll Call Vote in House

RCLatestYea_House: Number of Yea in Latest Roll Call Vote in House

RCLatestNay_House: Number of Nay in Latest Roll Call Vote in House

RCTimes_Senate: Number of Roll Call Voting Opportunities in Senate

RCLatestDate_Senate: Date of Latest Roll Call in Senate

RCLatestText_Senate: Content of Latest Roll Call Vote in Senate

RCLatestNumber_Senate: ID Number of Latest Roll Call Vote in Senate

RCLatestYea_Senate: Number of Yea in Latest Roll Call Vote in Senate

RCLatestNay_Senate: Number of Nay in Latest Roll Call Vote in Senate

If latestonly==FALSE, A data.frame object containing following variables:

Chamber: Chamber where the Action is Placed.

Date: Date of Action

Time: Time (in Day) of Action

Detail: Detailed Content of Action

Overview: Summarized Overview Content of Action (If Available)

Amendment: 1 if Actions is related to amendment of the bill

VoteType: Type of Vote (if any decision is made)

rollnumber: Roll Call Number if Action is Roll Call Vote

rollYea: Roll Call Yea Votes if Action is Roll Call Vote

rollNay: Roll Call Nay Votes if Action is Roll Call Vote

See also

Examples

tgturl <- "https://www.congress.gov/bill/115th-congress/house-bill/2"
bill2 <- scrapeBill(tgturl)
collectAction(bill2)

#>   ActLatestDate_Overview ActLatestText_Overview ActLatestDate_All
#> 1             12/20/2018             Became Law        12/20/2018
#>                ActLatestText_All VoteTimes_House VoteLatestDate_House
#> 1 Became Public Law No: 115-334.               0                 <NA>
#>   VoteLatestType_House VoteLatestText_House VoteTimes_Senate
#> 1                 <NA>                 <NA>                0
#>   VoteLatestDate_Senate VoteLatestType_Senate VoteLatestText_Senate
#> 1                  <NA>                  <NA>                  <NA>
#>   VoteTimes_House.1 RCLatestDate_House RCLatestText_House RCLatestNumber_House
#> 1                 0               <NA>               <NA>                   NA
#>   RCLatestYea_House RCLatestNay_House RCTimes_Senate RCLatestDate_Senate
#> 1                NA                NA              0                <NA>
#>   RCLatestText_Senate RCLatestNumber_Senate RCLatestYea_Senate
#> 1                <NA>                    NA                 NA
#>   RCLatestNay_Senate
#> 1                 NA